Refactor music controller into its own package
The music controller was a single large file, so it could not carry a
strings.json for server-owned name/description localization like the other
folder-based controllers (cache, players, ...).

- move controllers/music.py -> controllers/music/ package (controller.py)
- move controllers/media/ -> controllers/music/media/ (sub-controllers of music)
- split out constants.py, database.py (MusicDatabaseSetupMixin), migrations.py and helpers.py
- add core.music strings.json (manifest + config_entries) + README.md, moving the
  reset_db config-entry strings out of the common strings.json
- add tests for the search-result sort helper, db-setup wiring and migrations

The public import path music_assistant.controllers.music (and MusicController)
is unchanged.
