+++ music_assistant/controllers/music/strings.json
@@ -1,0 +1,15 @@
+{
+  "manifest": {
+    "name": "Music",
+    "description": "Configuration of the core music (library) management module"
+  },
+  "config_entries": {
+    "reset_db": {
+      "label": "Reset library database",
+      "description": "This will issue a full reset of the library database and trigger a full sync. Only use this option as a last resort if you are seeing issues with the library database."
+    },
+    "reset_db_result": {
+      "label": "The database has been reset."
+    }
+  }
+}
