+++ music_assistant/translations/en.json
@@ -160,9 +160,6 @@
   "common.config_entries.protocol_enable.label": "Enable",
   "common.config_entries.publish_ip.description": "This IP address is communicated to players where to find this server.\nMake sure that this IP can be reached by players on the local network, otherwise audio streaming will not work.",
   "common.config_entries.publish_ip.label": "Published IP address",
-  "common.config_entries.reset_db.description": "This will issue a full reset of the library database and trigger a full sync. Only use this option as a last resort if you are seeing issues with the library database.",
-  "common.config_entries.reset_db.label": "Reset library database",
-  "common.config_entries.reset_db_result.label": "The database has been reset.",
   "common.config_entries.sample_rates.description": "The sample rates (and bit depths) supported by this player.\nContent with unsupported sample rates will be automatically resampled.",
   "common.config_entries.sample_rates.label": "Sample rates supported by this player",
   "common.config_entries.sample_rates.options.176400||16": "176.4kHz / 16 bits",
@@ -445,6 +442,11 @@
   "core.discovery.config_entries.upnp_network_scan.label": "Allow network scan for UPnP discovery",
   "core.discovery.manifest.description": "Handles mDNS/Zeroconf, SSDP/UPnP discovery and Music Assistant network broadcast.",
   "core.discovery.manifest.name": "Discovery",
+  "core.music.config_entries.reset_db.description": "This will issue a full reset of the library database and trigger a full sync. Only use this option as a last resort if you are seeing issues with the library database.",
+  "core.music.config_entries.reset_db.label": "Reset library database",
+  "core.music.config_entries.reset_db_result.label": "The database has been reset.",
+  "core.music.manifest.description": "Configuration of the core music (library) management module",
+  "core.music.manifest.name": "Music",
   "core.player_queues.config_entries.default_enqueue_option_album.description": "Define the default enqueue action for this mediatype.",
   "core.player_queues.config_entries.default_enqueue_option_album.label": "Default enqueue option for Album item(s).",
   "core.player_queues.config_entries.default_enqueue_option_album.options.play": "Play (and keep queue)",
